Eus upgrade automation (#124)
* Enable eus upgrade option


* tatusAutomate EUS to EUS upgrade


# Conflicts:
#	README.md

# Conflicts:
#	README.md

* Allow user to choose target version for eus upgrade

Until now, only 4.Y+2.0 target version was allowed for EUS upgrade.
With direct upgrade paths allowed,
the option to choose target cnv version is required.


* Add/Fix typing of util functions


* Move workloadupdatestrategy to outer scope


---------

+++ tests/install_upgrade_operators/utils.py
@@ -4,6 +4,7 @@ import logging
 import re
 
 from benedict import benedict
+from kubernetes.dynamic import DynamicClient
 from kubernetes.dynamic.exceptions import ConflictError, ResourceNotFoundError
 from ocp_resources.installplan import InstallPlan
 from ocp_resources.network_addons_config import NetworkAddonsConfig
@@ -17,6 +18,7 @@ from utilities.constants import (
     PRODUCTION_CATALOG_SOURCE,
     TIMEOUT_1MIN,
     TIMEOUT_5SEC,
+    TIMEOUT_10SEC,
     TIMEOUT_30MIN,
     TIMEOUT_40MIN,
 )
@@ -29,7 +31,7 @@ def wait_for_operator_condition(dyn_client, hco_namespace, name, upgradable):
     LOGGER.info(f"Wait for the operator condition. Name:{name} Upgradable:{upgradable}")
     samples = TimeoutSampler(
         wait_timeout=TIMEOUT_30MIN,
-        sleep=1,
+        sleep=TIMEOUT_10SEC,
         func=OperatorCondition.get,
         dyn_client=dyn_client,
         namespace=hco_namespace,
@@ -53,12 +55,20 @@ def wait_for_operator_condition(dyn_client, hco_namespace, name, upgradable):
         raise
 
 
-def wait_for_install_plan(dyn_client, hco_namespace, hco_target_version, is_production_source):
+def wait_for_install_plan(
+    dyn_client: DynamicClient,
+    hco_namespace: str,
+    hco_target_version: str,
+    is_production_source: bool,
+) -> InstallPlan:
     install_plan_sampler = TimeoutSampler(
         wait_timeout=TIMEOUT_40MIN,
-        sleep=1,
+        sleep=TIMEOUT_10SEC,
         func=InstallPlan.get,
-        exceptions_dict={ConflictError: []},  # Ignore ConflictError during install plan reconciliation
+        exceptions_dict={
+            ConflictError: [],
+            ResourceNotFoundError: [],
+        },  # Ignore ConflictError during install plan reconciliation
         dyn_client=dyn_client,
         hco_namespace=hco_namespace,
         hco_target_version=hco_target_version,
@@ -80,25 +90,27 @@ def wait_for_install_plan(dyn_client, hco_namespace, hco_target_version, is_prod
                 # upgrade cnv using custom catalogsource, to a specified version. Approving install plan associated
                 # with the production catalogsource would also lead to failure as production catalogsource has been
                 # disabled at this point.
-                ip_instance = ip.instance
-                if not is_production_source:
+                if ip.exists:
+                    ip_instance = ip.instance
+                    if not is_production_source:
+                        if (
+                            not ip_instance.spec.approved
+                            and ip_instance.status
+                            and ip_instance.status.bundleLookups[0].get("catalogSourceRef").get("name")
+                            == PRODUCTION_CATALOG_SOURCE
+                        ):
+                            ip.clean_up()
+                            continue
                     if (
-                        not ip_instance.spec.approved
-                        and ip_instance.status
-                        and ip_instance.status.bundleLookups[0].get("catalogSourceRef").get("name")
-                        == PRODUCTION_CATALOG_SOURCE
+                        hco_target_version == ip_instance.spec.clusterServiceVersionNames[0]
+                        and ip.name == install_plan_name_in_subscription
                     ):
-                        ip.delete(wait=True)
-                        continue
-                if (
-                    hco_target_version == ip_instance.spec.clusterServiceVersionNames[0]
-                    and ip.name == install_plan_name_in_subscription
-                ):
-                    return ip
-                LOGGER.info(
-                    f"Subscription: {subscription.name}, is associated with install plan:"
-                    f" {install_plan_name_in_subscription}"
-                )
+                        return ip
+                    LOGGER.info(
+                        f"Subscription: {subscription.name}, is associated with install plan:"
+                        f" {install_plan_name_in_subscription}"
+                    )
+
     except TimeoutExpiredError:
         LOGGER.error(
             f"timeout waiting for target install plan: version={hco_target_version}, "
